Update info, rename classes and move stuff
