+++ .env.example
@@ -1,6 +1,6 @@
 KOL_USER='Cagebot'
 KOL_PASS='Cagebot P4ssw0rD'
-MAINTAIN_ADVENTURES='80'
+MAINTAIN_ADVENTURES='100'
 OPEN_EVERYTHING='true/false'
 ONLY_OPEN_WHEN_ADVS_ABOVE='80'
 WHITEBOARD_CAGED='${name} (#${id}) is CAGED in Hobopolis'
